Finish ansible rewrite
